@@ -34,20 +34,28 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
-  firebase_database: ^10.5.7
-  flutter_lints: ^5.0.0
-  cloud_firestore: ^4.15.4
-  geoflutterfire2: ^2.3.15
-  google_maps_flutter: ^2.6.0
-  geolocator: ^11.0.0
-  firebase_core: ^2.32.0
-  firebase_auth: ^4.11.0
-  google_sign_in: ^6.3.0
+  # Firebase packages - compatible versions
+  firebase_core: ^3.6.0
+  firebase_auth: ^5.3.1
+  firebase_database: ^11.1.4
+  cloud_firestore: ^5.4.4
+
+  # Google Sign In
+  google_sign_in: ^6.2.1
+
+  # Maps and Location
+  google_maps_flutter: ^2.9.0
+  geolocator: ^13.0.1
+
+  # Using maintained fork instead of geoflutterfire
+  geoflutterfire_plus: ^0.0.3
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+  flutter_lints: ^5.0.0
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
