@@ -1,6 +1,3 @@
-import 'package:aidkriya_walker/find_walker_screen.dart';
-import 'package:aidkriya_walker/setup_flow_screen.dart';
-import 'package:aidkriya_walker/sign_in_screen.dart';
 import 'package:aidkriya_walker/sign_up_screen.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
@@ -9,9 +6,7 @@ import 'firebase_options.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  await Firebase.initializeApp(
-    options: DefaultFirebaseOptions.currentPlatform,
-  );
+  await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
   runApp(const MyApp());
 }
 
